search activity now displays text information from yelp
Search activity now populates the list view with not only the restaurant
icons but also the restaurant name, the restaurant display
address([street address], [city], [state]), and the rating from yelp
multiplied by 2. Still left to do: distance and ranking based on
distance. It is possible to implement ranking based on rating.